+++ test/smee_view/views/entities_test.exs
@@ -50,6 +50,49 @@ defmodule SmeeViewEntitiesTest do
               = ThisView.view_one([@src_md, @src_entity])
     end
 
+    #########
+
+    test "entity_id_hash has is set from the original Entity" do
+      %Aspects.Entity{entity_id_hash: "c0045678aa1b1e04e85d412f428ea95d2f627255"} = ThisView.view_one(@src_entity)
+    end
+
+    test "id is set from the original Entity" do
+      %Aspects.Entity{id: "e123456"} = struct(ThisView.view_one(@src_entity), %{id: "e123456"})
+      %Aspects.Entity{id: nil} = ThisView.view_one(@src_entity)
+    end
+
+    test "valid_until is set from the original Entity, as a DateTime" do
+      %Aspects.Entity{valid_until: ~U[2022-01-01 16:22:44.834Z]} = ThisView.view_one(@src_entity)
+    end
+
+    test "modified_at is set from the original Entity, as a DateTime" do
+      %Aspects.Entity{modified_at: ~U[2023-04-26 10:36:51Z]} = struct(ThisView.view_one(@src_entity), %{modified_at: ~U[2023-04-26 10:36:51Z]})
+    end
+
+    test "downloaded_at is set from the original Entity, as a DateTime" do
+      %Aspects.Entity{downloaded_at: ~U[2024-03-04 11:10:40.308331Z]} = struct(ThisView.view_one(@src_entity), %{downloaded_at: ~U[2024-03-04 11:10:40.308331Z]})
+    end
+
+    test "priority is set from the original Entity, as an integer" do
+      %Aspects.Entity{priority: 8} = struct(ThisView.view_one(@src_entity), %{priority: 8})
+    end
+
+    test "trustiness is set from the original Entity, as a float" do
+      %Aspects.Entity{trustiness: 0.8} = struct(ThisView.view_one(@src_entity), %{trustiness: 0.8})
+    end
+
+    test "tags is set from the original Entity, as a list of binaries (or empty)" do
+      %Aspects.Entity{tags: ["bar", "foo"]} = struct(ThisView.view_one(@src_entity), %{tags: ["bar", "foo"]})
+    end
+
+    test "label is set from the original Entity" do
+      %Aspects.Entity{label: "A label"} = struct(ThisView.view_one(@src_entity), %{label: "A label"})
+    end
+
+    test "entity_id from XML must match the uri field in the originating entity" do
+      assert_raise RuntimeError, fn -> ThisView.view_one(struct(@src_entity, %{uri: "http://example.com"})) end
+    end
+
   end
 
   describe "prism/2" do