+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2023-02-
+## [0.2.0] - 2023-03-
 
 ### Breaking Changes
 
 
 ### New Features
+- Entity Aspects now contain the extra data from Smee Entities, not just XML-derived data, so tags, download times, etc
+  can be accessed all from the same aspect struct. 
 
 ### Improvements
 - Now uses the improved XML processing in Smee v0.4.0, doubling the speed